backend-api/requirements.txt
fastapi
uvicorn[standard]
groq
python-dotenv
regex
googlesearch-python
beautifulsoup4